allow to move from one pointer to another
