+++ src/main/java/com/github/lopoha/coboltruffle/parser/CobolNodeFactory.java
@@ -5,6 +5,7 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
+import com.github.lopoha.coboltruffle.parser.nodes.CobolMoveNode;
 import com.oracle.truffle.api.RootCallTarget;
 import com.oracle.truffle.api.Truffle;
 import com.oracle.truffle.api.frame.FrameDescriptor;
